requirements.txt
requests
azure-storage